requirements.txt
Flask_SQLAlchemy==2.5.1
Flask==1.1.2
spotipy==2.18.0
Werkzeug==1.0.1
SQLAlchemy==1.3.18
cloudinary==1.26.0
Flask_Login==0.5.0
requests==2.26.0
beautifulsoup4==4.9.3
python-dotenv==0.19.0
uvicorn
gunicorn==19.9.0
psycopg2-binary
